API Updates

+++ server/src/graphql-schemas/videoNotes.js
@@ -8,6 +8,7 @@ const { DBRef } = require('mongodb');
 const mongo = require('mongodb');
 const studyHelper = require('../func/study');
 const videoHelper = require('../func/videoNote')
+const permissions = require('../func/permissionsChecker')
 
 const typeDefs = gql`
     type Note {
@@ -151,7 +152,9 @@ const resolvers = {
                     const NotesCollection = database.getDb().collection('notes');
                     var n_id = new mongo.ObjectID(arg.videoNotesID);
                     const currNotes = await NotesCollection.findOne({ "_id": n_id })
-
+                    if (!await permissions.permissionChecker(ctx, currNotes.study.oid, "read")) {
+                        throw ForbiddenError("Invalid Permissions")
+                    }
                     // Returns item if it exists or throws error
                     if (currNotes) {
                         return await videoHelper.formVideoNote(currNotes)
@@ -185,6 +188,9 @@ const resolvers = {
                     const NotesCollection = database.getDb().collection('notes');
                     var s_id = new mongo.ObjectID(arg.studyID);
                     const currNotes = await NotesCollection.find({ study: DBRef("study", s_id) }).toArray()
+                    if (ctx.user.Level < 2) {
+                        throw new ForbiddenError("Invalid Permissions")
+                    }
                     if (currNotes) {
                         var noteList = []
                         for (let x in currNotes) {
@@ -229,6 +235,9 @@ const resolvers = {
                 const study = await studyHelper.getStudy(study_oid)
                 if (!study) {
                     throw new IdError("Invalid studyID")
+                }
+                if (!await permissions.permissionChecker(ctx, study.oid, "create")) {
+                    throw ForbiddenError("Invalid Permissions")
                 }
                 try {
                     // Build the arrays for any possible videos and notes
@@ -293,6 +302,9 @@ const resolvers = {
                 if (!currNotes) {
                     throw new IdError("Invalid videoNotesID")
                 }
+                if (!await permissions.permissionChecker(ctx, currNotes.study.oid, "create")) {
+                    throw ForbiddenError("Invalid Permissions")
+                }
                 try {
                     const response = await NotesCollection.updateOne(
                         { "_id": n_id },
@@ -335,6 +347,9 @@ const resolvers = {
                 var currNotes = await NotesCollection.findOne({ "_id": n_id })
                 if (!currNotes) {
                     throw new IdError("Invalid videoNotesID")
+                }
+                if (!await permissions.permissionChecker(ctx, currNotes.study.oid, "create")) {
+                    throw ForbiddenError("Invalid Permissions")
                 }
                 try {
                     const response = await NotesCollection.updateOne(
@@ -380,6 +395,9 @@ const resolvers = {
                 if (!currNotes) {
                     throw new IdError("Invalid videoNotesID")
                 }
+                if (!await permissions.permissionChecker(ctx, currNotes.study.oid, "edit")) {
+                    throw ForbiddenError("Invalid Permissions")
+                }
                 try {
                     updateField = {}
                     if ('title' in arg) {
@@ -415,6 +433,9 @@ const resolvers = {
                 var currNotes = await NotesCollection.findOne({ "_id": n_id })
                 if (!currNotes) {
                     throw new IdError("Invalid videoNotesID")
+                }
+                if (!await permissions.permissionChecker(ctx, currNotes.study.oid, "edit")) {
+                    throw ForbiddenError("Invalid Permissions")
                 }
                 var existCheck = false
                 const note_id = new mongo.ObjectID(arg.noteID);
@@ -469,6 +490,9 @@ const resolvers = {
                 var currNotes = await NotesCollection.findOne({ "_id": n_id })
                 if (!currNotes) {
                     throw new IdError("Invalid videoNotesID")
+                }
+                if (!await permissions.permissionChecker(ctx, currNotes.study.oid, "edit")) {
+                    throw ForbiddenError("Invalid Permissions")
                 }
                 var existCheck = false
                 const video_id = new mongo.ObjectID(arg.videoID);
@@ -530,6 +554,9 @@ const resolvers = {
                 if (!currNotes) {
                     throw new IdError("Invalid videoNotesID")
                 }
+                if (!await permissions.permissionChecker(ctx, currNotes.study.oid, "delete")) {
+                    throw ForbiddenError("Invalid Permissions")
+                }
                 await NotesCollection.deleteOne({ "_id": n_id })
                 return await videoHelper.formVideoNote(currNotes)
             } else {
@@ -544,6 +571,12 @@ const resolvers = {
                 const NotesCollection = database.getDb().collection('notes');
                 const n_id = new mongo.ObjectID(arg.videoNotesID);
                 var currNotes = await NotesCollection.findOne({ "_id": n_id })
+                if (!currNotes) {
+                    throw new IdError("Invalid videoNotesID")
+                }
+                if (!await permissions.permissionChecker(ctx, currNotes.study.oid, "delete")) {
+                    throw ForbiddenError("Invalid Permissions")
+                }
                 var existCheck = false
                 const video_id = new mongo.ObjectID(arg.videoID);
                 for (let x in currNotes.videos) {
@@ -553,9 +586,6 @@ const resolvers = {
                 }
                 if (!existCheck) {
                     throw new IdError("Invalid videoID")
-                }
-                if (!currNotes) {
-                    throw new IdError("Invalid videoNotesID")
                 }
                 await NotesCollection.updateOne(
                     { "_id": n_id },
@@ -575,12 +605,18 @@ const resolvers = {
                 )
             }
         },
-        
+
         deleteNotefromVideoNote: async (parent, arg, ctx, info) => {
             if (ctx.auth) {
                 const NotesCollection = database.getDb().collection('notes');
                 const n_id = new mongo.ObjectID(arg.videoNotesID);
                 var currNotes = await NotesCollection.findOne({ "_id": n_id })
+                if (!currNotes) {
+                    throw new IdError("Invalid videoNotesID")
+                }
+                if (!await permissions.permissionChecker(ctx, currNotes.study.oid, "delete")) {
+                    throw ForbiddenError("Invalid Permissions")
+                }
                 var existCheck = false
                 const note_id = new mongo.ObjectID(arg.noteID);
                 for (let x in currNotes.notes) {
@@ -590,9 +626,6 @@ const resolvers = {
                 }
                 if (!existCheck) {
                     throw new IdError("Invalid noteID")
-                }
-                if (!currNotes) {
-                    throw new IdError("Invalid videoNotesID")
                 }
                 await NotesCollection.updateOne(
                     { "_id": n_id },
